prott5 model

+++ src/training/focal_loss.py
@@ -15,7 +15,7 @@ class FocalLoss(nn.Module):
 
         # compute cross entropy loss using F.cross_entropy with the weights
         # set reduction=None to get the loss for each sample
-        ce_loss = F.cross_entropy(input=input, target=target, weight=self.alpha, reduction="none")
+        ce_loss = F.cross_entropy(input=input.to(torch.float), target=target.to(torch.float), weight=self.alpha, reduction="none")
 
         # select the probabilities of the target class
         p = p.gather(index=target.unsqueeze(1), dim=1).squeeze()
